+++ test/webpack.config.js
@@ -28,5 +28,7 @@ module.exports = env => {
     rootDir: `${__dirname}/..`,
     getAliases
   });
+  // console.log('webpack env', JSON.stringify(env));
+  // console.log('webpack config', JSON.stringify(config, null, 2));
   return config;
 };
